Merge branch 'develop' of codeup.aliyun.com:fly/Link/Link_Service into feature/ycx

        https://codeup.aliyun.com/fly/Link/Link_Service/merge_request/10122021

+++ restful/user/internal/types/types.go
@@ -58,5 +58,5 @@ type UserUpdateRemarkRequest struct {
 type UserFlowedRequest struct {
 	BeId    uint64 `json:"beId"`
 	Message string `json:"message"`
-	Type    uint64 `json:"type"`
+	Type    uint32 `json:"type"`
 }
